Peer ID -> index in results.
Changed the output to show the peer's index, rather than the peer's ID. Peer
ID's change between consecutive runs, but their indices don't. This makes it
easier to track meaningful changes in the diffs of results files. There may be
a small subset of cases where having the ID information can help in debugging
(e.g. in case of issues involving IPTB's handling of Docker containers), but
these are probably so rare that it's not a primary concern (and this debugging
can happen in other ways).