Store Bitswap ledgers over time.
While a node downloads a file (using `ipfs get`), we now continually get that
peer's Bitswap ledgers. This gives a rough semblance of a time series that
can act as a first step toward analyzing the dynamics of the interactions over
time. The ledgers are stored in files alongside the aggregate results.

Other changes:

-   The results file argument, `-o`, has been replaced by a directory argument,
    `-d`, that places all of the results (aggregate, as well as ledger files
    for each peer) in that directory.
-   The general results file (that has all of the data saved by previous
    versions of the test script) is now called `aggregate` (rather than being
    named by the `-o` argument).

License: MIT
Signed-off-by: David Grisham <dgrisham@mines.edu>